% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piamModelTests-package.R
\docType{package}
\name{piamModelTests-package}
\alias{piamModelTests}
\alias{piamModelTests-package}
\title{piamModelTests: piamModelTests Tools}
\description{
A collection of R tools provided by the Integrated Assessment Modeling Consortium (IAMC) for data analysis and diagnostics.
}
\author{
\strong{Maintainer}: Jan Philipp Dietrich \email{dietrich@pik-potsdam.de}

Authors:
\itemize{
  \item Cornelia Auer
  \item Anastasis Giannousakis
  \item Christoph Bertram
  \item Falk Benke
  \item Florian Humpenoeder
  \item Lavinia Baumstark
}

}
